Add a macro for testing my eval against r5rs eval
This macro just asserts that the result of my eval is equal to the
result of the r5rs builtin eval. Most expressions should be identical
since this is a scheme implementation, therefore this simplifies
testing. However, this macro won't work for cases where r5rs does not
specify behavior - for example what a `define` expression evaluates to.