Add boolean? to self-evaluating? predicate
Subtle issues arise when following the book's recommendation to return
the symbol 'false or 'true as boolean values. In fact, the book installs
those symbols not as a value but as a variable whose value is the
underlying scheme's true/false. Therefore, when certain expressions
return 'false or 'true, they are in fact returning an unevaluated
variable's name rather than the actual value of true or false. To avoid
all this, I just made booleans self-evaluating and use the underlying
scheme's booleans to eval and return values rather than variables. I
still kept the book's initial environment with true/false defined
variables.